requirements.txt
aiogram>=3.0.0
sqlalchemy>=2.0.0
aiosqlite
asyncpg
python-dotenv
yt-dlp